Do not treat mouseout as mouseup when tracking document
Touchy treats the mouseout event as a mouseup event except when the
related target of the event is the elemenet being tracked or one of its
descendants. However when tracking 'document' and the mouse goes out of
the window bounds, the mouseup event shouldn't be fired because it's
still possible to track the mouse movement.